pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "text-to-sql"
version = "0.1.0"
description = "Text-to-SQL pipeline with multi-provider LLM support, MCP tools, and human-in-the-loop approval"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "mcp>=1.0.0",
    "litellm>=1.60.0",
    "anthropic>=0.45.0",
    "openai>=2.21.0",
    "google-cloud-bigquery>=3.40.1",
    "sqlalchemy[asyncio]>=2.0.46",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.21.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.8.0",
    "python-dotenv>=1.1.0",
    "structlog>=25.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "coverage>=7.6.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/text_to_sql"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
